Correcciones a la navegacion del proyecto
