+++ TPWinForm_Equipo18/Site.Master
@@ -42,14 +42,13 @@
 
         <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-dark bg-dark">
             <div class="container">
-                <a class="navbar-brand" runat="server" href="~/Default">TP 3 CARRITO</a>
+                <a class="navbar-brand" runat="server" href="~/Default.aspx">ListadoArticulos</a>
                 <button type="button" class="navbar-toggler" data-bs-toggle="collapse" data-bs-target=".navbar-collapse" title="Alternar navegación" aria-controls="navbarSupportedContent"
                     aria-expanded="false" aria-label="Toggle navigation">
                     <span class="navbar-toggler-icon"></span>
                 </button>
                 <div class="collapse navbar-collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
-                        <li class="nav-item"><a class="nav-link" runat="server" href="~/ListadoArticulos">ListadoArticulos</a></li
                         <li class="nav-item">
                             <a class="nav-link btn btn-success" runat="server" href="~/Carrito">
                                 <svg width="16" height="16" fill="currentColor" class="bi bi-cart-check" viewBox="0 0 16 16">
